drm/i915: Watchdog timeout: Blindly trust watchdog timeout for reset?
XXX: What to do when the watchdog irq fired twice but our hangcheck
logic thinks the engine is not hung? For example, what if the
active-head moved since the irq handler?

One option is to just ignore the watchdog, if the engine is really hung,
then the driver will detect the hang by itself later on (I'm inclined to
this).

But the other option is to blindly trust the HW, which is what this patch
does...

Signed-off-by: Michel Thierry <michel.thierry@intel.com>